requirements.txt
amqp==5.0.6
astroid==2.4.2
autopep8==1.5.4
billiard==3.6.4.0
celery==5.0.5
certifi==2019.3.9
chardet==4.0.0
click==7.1.2
click-didyoumean==0.0.3
click-plugins==1.1.1
click-repl==0.1.6
colorama==0.4.3
coreapi==2.3.3
coreschema==0.0.4
Django==2.2.2
django-cors-headers==3.7.0
django-rest-auth==0.9.5
django-rest-swagger==2.2.0
django-widget-tweaks==1.4.8
djangorestframework==3.12.1
idna==2.10
image==1.5.27
importlib-metadata==4.0.1
isort==4.3.21
itypes==1.2.0
Jinja2==3.0.0
kombu==5.0.2
lazy-object-proxy==1.4.3
MarkupSafe==2.0.0
mccabe==0.6.1
numpy==1.16.4
openapi-codec==1.3.2
opencv-python==4.1.0.25
Pillow==6.0.0
pipenv==2018.11.26
prompt-toolkit==3.0.18
pycodestyle==2.6.0
pylint==2.5.3
PyMySQL==1.0.2
pytesseract==0.2.6
pytz==2019.1
redis==3.5.3
requests==2.25.1
simplejson==3.17.2
six==1.14.0
sqlparse==0.3.0
toml==0.10.1
typed-ast==1.4.1
typing-extensions==3.10.0.0
uritemplate==3.0.1
urllib3==1.26.4
vine==5.0.0
virtualenv==16.6.0
virtualenv-clone==0.5.3
wcwidth==0.2.5
wrapt==1.11.2
zipp==3.4.1
